Re-implement code that finds local interface addresses (vhosts).
The old method was to shell out to various system utilities (ifconfig, 
netstat etc) and parse the result.  This was vulnerable to a symlink 
attack, and was obviously pretty crufty anyhow.

The new method assumes we have getifaddrs() on IPv6-capable systems, and
the SIOCGIFCONF ioctl on IPv4 systems.  /proc/net/if_inet6 is used to
supplement getifaddrs() to cater for Linux systems with older glibc that
didn't return IPv6 addresses from getifaddrs().

I've tested this on Linux, FreeBSD and Cygwin and It Seems To Work...

This fixes the vulnerability CVE-2007-5839 (cve.mitre.org).


git-svn-id: https://bitchx.svn.sourceforge.net/svnroot/bitchx/branches/ircii-pana-1.1@11 13b04d17-f746-0410-82c6-800466cd88b0